feat: add options to consolidate the window size
When constructing a compatible reader with get_compatible the window size was previously set to the minimum of the original readers and the new readers default window size for both readers. Now there are options to only use the original readers window size or to let the new reader use its default one.